added interactions

+++ src/routes/interactions.js
@@ -5,13 +5,19 @@ import { conFn } from '../mysql.js';
 export async function get({ params }) {
 	let con = await conFn();
 
-	const users = await con.query('SELECT * FROM website_user;').then(function ([rows, fields]) {
-		return rows;
-	});
-	if (users) {
+	const interactions = await con
+		.query(
+			`SELECT a.first_name,last_name, Date_Format(i.interaction_date,"%m/%d/%Y") as interaction_date, i.notes
+	FROM interactions as i JOIN high_school_athlete as a
+		ON a.athlete_id = i.athlete_id;`
+		)
+		.then(function ([rows, fields]) {
+			return rows;
+		});
+	if (interactions) {
 		return {
 			status: 200,
-			body: { users }
+			body: { interactions }
 		};
 	}
 }
@@ -48,7 +54,7 @@ export async function put({ request }) {
 	const user_id = form.get('user_id');
 	con.query(
 		'UPDATE website_user SET first_name = ?,  last_name = ?, email = ?, password = ? WHERE user_id = ?',
-		[first_name, last_name, email, password,user_id]
+		[first_name, last_name, email, password, user_id]
 	);
 	return {
 		body: {
@@ -67,10 +73,7 @@ export async function del({ request }) {
 	const form = await request.formData();
 
 	const user_id = form.get('user_id');
-	con.query(
-		'DELETE FROM website_user WHERE user_id = ?',
-		[user_id]
-	);
+	con.query('DELETE FROM website_user WHERE user_id = ?', [user_id]);
 	return {
 		body: {
 			first_name: form.get('first_name'),
@@ -81,4 +84,3 @@ export async function del({ request }) {
 		}
 	};
 }
-
